+++ light-std/src/lstd/io/string_writer.h
@@ -8,7 +8,7 @@ LSTD_BEGIN_NAMESPACE
 
 namespace io {
 
-void string_writer_write(writer *w, const char *data, size_t count);
+void string_writer_write(writer *w, const char *data, s64 count);
 
 struct string_writer : writer {
     string *Str;
@@ -16,12 +16,12 @@ struct string_writer : writer {
     string_writer(string *str) : writer(string_writer_write, writer_flush_do_nothing), Str(str) {}
 };
 
-inline void string_writer_write(writer *w, const char *data, size_t count) {
+inline void string_writer_write(writer *w, const char *data, s64 count) {
     auto *sw = (string_writer *) w;
     sw->Str->append_pointer_and_size(data, count);
 }
 
-void string_builder_writer_write(writer *w, const char *data, size_t count);
+void string_builder_writer_write(writer *w, const char *data, s64 count);
 
 struct string_builder_writer : writer {
     string_builder Builder;
@@ -29,7 +29,7 @@ struct string_builder_writer : writer {
     string_builder_writer() : writer(string_builder_writer_write, writer_flush_do_nothing) {}
 };
 
-inline void string_builder_writer_write(writer *w, const char *data, size_t count) {
+inline void string_builder_writer_write(writer *w, const char *data, s64 count) {
     auto *sw = (string_builder_writer *) w;
     sw->Builder.append_pointer_and_size(data, count);
 }
