+++ light-std/src/lstd/io/writer.h
@@ -19,21 +19,21 @@ inline void writer_flush_do_nothing(writer *) {}
 // in this class is implemented by calling those functions.
 // By default _FlushFunction_ points to a stub that does nothing.
 struct writer : non_copyable, non_movable, non_assignable {
-    using write_t = void (*)(writer *w, const char *data, size_t count);
+    using write_t = void (*)(writer *w, const char *data, s64 count);
     using flush_t = void (*)(writer *w);
 
     write_t WriteFunction = null;
     flush_t FlushFunction = writer_flush_do_nothing;
 
     char *Buffer = null, *Current = null;
-    size_t BufferSize = 0, Available = 0;
+    s64 BufferSize = 0, Available = 0;
 
     writer() = default;
     writer(write_t writeFunction, flush_t flushFunction) : WriteFunction(writeFunction), FlushFunction(flushFunction) {}
 
     void write(array_view<char> data) { WriteFunction(this, data.begin(), data.size()); }
     void write(const char *data) { WriteFunction(this, data, c_string_length(data)); }
-    void write(const char *data, size_t count) { WriteFunction(this, data, count); }
+    void write(const char *data, s64 count) { WriteFunction(this, data, count); }
     void write(string str) { WriteFunction(this, str.Data, str.ByteLength); }
 
     void write(char32_t cp) {
