+++ light-std/src/lstd/memory/pixel_buffer.cpp
@@ -23,7 +23,7 @@ pixel_buffer::pixel_buffer(file::path path, bool flipVertically, pixel_format fo
     // // @Speed Killmenowkillmefastkillmenowkillmefast
     // s32 w, h, n;
     // u8 *loaded = stbi_load_from_memory((const u8 *) data.Data, (s32) data.ByteLength, &w, &h, &n, (s32) format);
-    // size_t allocated = ((allocation_header *) loaded - 1)->Size;
+    // s64 allocated = ((allocation_header *) loaded - 1)->Size;
     //
     // loaded = (u8 *) allocator::reallocate(loaded, allocated + POINTER_SIZE);
     // copy_memory(loaded + POINTER_SIZE, loaded, w * h * n);
@@ -50,7 +50,7 @@ void pixel_buffer::release() {
 
 pixel_buffer *clone(pixel_buffer *dest, pixel_buffer src) {
     *dest = src;
-    size_t size = src.Width * src.Height * src.BPP;
+    s64 size = src.Width * src.Height * src.BPP;
     dest->Pixels = new u8[size];
     copy_memory(dest->Pixels, src.Pixels, size);
     return dest;
